refactoring code, separated suggestion modal as another component
